gcc: do not use an old cross-compiler build a new native compiler
Closes #244871

This commit checks for the following situation: we are
cross-building a native compiler (build!=(host==target)) using a
cross-compiler ((build==host)!=target) which is *older than* the
compiler being built.  We check both the cross-compiler itself as
well as the cross-compiler which was used to build the copy of
libgcc against which the target libc is linked.

This configuration often fails (see
https://github.com/NixOS/nixpkgs/issues/244871 for one example).

If you need to cross-build a newer-than-default-gcc-version native
compiler, you must first build a newer-than-default-gcc-version
cross-compiler.  Previously, the user had to figure out how to hack
in this upgrade on their own.  The code added by this commit inserts
the extra compiler build automatically.