fusuma: enable "essential" plugins
Specifically the following plugins:

    fusuma-plugin-appmatcher
    fusuma-plugin-keypress
    fusuma-plugin-sendkey
    fusuma-plugin-tap
    fusuma-plugin-wmctrl

I've not enabled the other plugins available on rubygems for the
following reasons:

  * remap : seems niche functionality and requires further configuration
            to grant the user access permissions to event devices

  * thumbsense : pulls in remap (see above)

  * touchscreen : I've no way of testing.  Note: enabling didn't appear
                  to cause any problems.

Ideally the plugin functionality would be made available as separate
pkgs, but that would require patching Fusuma to search outside of the
Gem directory.  Enabling this subset of packages for what appears to be
widely useful functionality seems a good option.