lib.systems.parse: allow kernel=windows with abi=gnu
`kernel=windows` is sort of a nixpkgs-ism; we parse triples
differently than gnu-config but then undo the difference when
unparsing in order to match its behavior.

However when platform objects are constructed internally they don't
have access to this hackery, so we need to permit direct
construction of gnu-config-invalid platform objects in order to
cover this case.