uboot: restrict rpi kernel size patch to rpi uboot builds
NixOS maintains a uboot patch for the raspberry pi which enables
larger kernels to be loaded. (background: [1,2])

Prior to this commit, this patch was applied to all callers of
buildUBoot, regardless of whether it was relevant for that specific
board. buildUBoot already accepts an argument `extraPatches' for
board-specific patches, so use this mechanism to patch the raspberry
pi uboot builds instead.

Without this change, passing a different `src' argument to
buildUBoot (ie: pointing at upstream uboot or at a board-specific
fork) becomes highly brittle as the patch will not usually apply
cleanly.

[1]: github.com/NixOS/nixpkgs/pull/139825
[2]: github.com/NixOS/nixpkgs/issues/97064

Signed-off-by: Nicholas Sielicki <nix@opensource.nslick.com>