pre-commit: fixing import bug for built-in hooks
Pre-commit's "system" language is a sort of escape hach from the
encapsulation which pre-commit typically provides.  It just runs
commands without thinking too hard about setting up or tearing down
environments.

But it is used for two things: user-provided hooks which reference
things in calling environment, and pre-commit-provided hooks which
live in the pre-commit package (I'm calling these "meta" hooks).

The challenge becomes: which PATH and PYTHONPATH variables do we expose
to pre-commit?  If it's the calling environment, the meta hooks break,
and if it's the packaged environment, the external-facing system hooks
break.

In order to make both of these work at the same time, this commit
separates 'system' into two languages: 'system', and 'meta'. System
hooks get access to the calling environment, and meta hooks get access
to the packaged environment.