qemu-utils: copy from qemu_kvm, not qemu
qemu_kvm is a much smaller build, so it's nicer if you're building
qemu-utils specifically.  None of the tools depend on the emulation
targets disabled in qemu_kvm.

The tools are copied, not linked, so we don't have to worry about
potentially depending on both qemu_kvm and qemu in a bigger
derivation, and qemu_kvm is already built by Hydra, so there's no
increase in Hydra workload by using the variant.