ghc: silence error on non-x86_64 with clang
When the stdenv is using a newer clang, it fails on non-x86_64 platforms
due to an unused `popcount16` function, which is treated as an error.

This only affects GHC 9.4.x, which I assume is due to a quirk of its
build system that GHC 8.8.x, 8.10.x, 9.2.x, and 9.6.x do not share.