python3Packages.cairo-lang: drop
The Python cairo-lang package was devised to be a set of tooling for the 0.x Cairo language for general
provable computations via zero knowledge STARK mechanisms.

It has been superseded by the Rust compiler for the Cairo language nowadays which is available
in nixpkgs.

We do not need to keep this EOL and unmaintained package.