Remove service result check for oci-containers; fixes #249332
The container would only be stopped when the service result was not "success". 
This would cause the monitoring process to be killed together with the rest of the cgroup.
Killing the monitoring process could cause container processes to crash and coredump.
No check is present in the systemd units gennerated by `podman systemd generate`. The check is not needed because the stop action is not run if the unit failed to start.