uboot: fix build on Allwinner sunxi platforms
U-boot currently fails to build on Allwinner sunxi platforms with
support for SCP firmware. Providing a SCP firmware is optional, however
u-boot emits an error if not explicitly disabled.  It is recommended to
pass "/dev/null" via the SCP environment variable to explicitly disable
the error[0].

```
binman: Filename 'scp.bin' not found in input path (.,.,./board/sunxi,arch/arm/dts) (cwd='/build/u-boot-2023.07.02')
make: *** [Makefile:1111: .binman_stamp] Error 1
```

The expected firmware for SCP is Crust[1].  It exists to implement forms
of deep sleep.  It is especially useful on devices with batteries, such
as the PinePhone, and handles shutdown/reboot on devices without a PMIC.

Crust is supported upstream in u-boot since 2021.01-rc1 and ARM
Trusted Firmware-A since v2.3.  The patches required for Crust to
co-exist with the Linux kernel however are not upstream.

As the SCP firmware is optional, and any previously building forms of
uboot for these boards would have been built without Crust, excluding it
restores functionality without unexpectedly removing a feature.
Further including it would potentially break booting with unpatched
kernels, I have not attempted to test this.

I do not have access to any of these devices.  This was found as I'm
starting to attempt to bootstrap the OrangePi Prime which contains an
Allwinner H5.

[0]: https://elixir.bootlin.com/u-boot/v2023.07.02/source/board/sunxi/README.sunxi64#L85
[1]: https://github.com/crust-firmware/crust