mailman: remove docutils input
In Nixpkgs, we currently have the strange situation where
pkgs.docutils (which was being used here) is on a different version
than pkgs.python3.pkgs.docutils (which is propagated from sphinx).
This led to a conflict here and a build failure.

Since we'd need to provide the same version as propagated by sphinx
anyway, let's just remove the explicit docutils check input and only
use the version propagated by sphinx, avoiding this kind of conflict
in future.

Fixes: 2312861130ec ("docutils: 0.19 -> 0.20.1")