nixos/modules: show warning and continue on self import
Before this PR, when a module recursively imports itself, the module
system would throw a stack overflow.
After this change, we show a warning and continue the evaluation.

To track the modules that we've already seen, we keep track of their
module keys in an attr set and at every recursive step we verify that
none of the keys of the new modules is already present in this set.
If any such key is found, we print a warning and ignore the recursive
import.

Fixes #252096