firefox wrapper: fix merging of policies files
If multiple extraPolicies or extraPoliciesFiles are given, the JSON
objects are merged using `jq` with the `a + b` operator, which simply
combines all the unique keys of `a`, `b` with the duplicated keys from
`b`, without recursion.
This strategy is completely wrong in this case: as policy files consists
of a single key, "policies", all that happens is that `b` takes over, in
other words:

    $(jq -s '.[0] + .[1]' a b) == $(cat b)

So there is no merging at all, the final policies.json file is simply
the last file in the list.

The `a * b` operation should be used instead, which performs the merge
by recursing in each key.