jimtcl: Fix cross compilation and allow disable of SDL
Jimtcl is dependency of OpenOCD that does not need SDL. For some
embedded systems it is beneficial to not pull in graphics related
dependencies to reduce the size. The second reason is that right now
SDL_gfx is not cross compiling and thus OpenOCD can't be cross compiled
either.

This also fixes cross compilation of Jimtcl itself. The attempt to cross
compile resulted in:

  No installed jimsh or tclsh, building local bootstrap jimsh0

The solution here is to add jimtcl to native build dependencies in case
of cross compilation.