frida.frida-gum: init at 16.1.8
This is kind of a hot mess due to upstream using various variously
cursed forks of dependencies. This includes, incidentally, a fork of
capstone.

This is structured as a top level frida attribute because the frida
project includes a whole bunch of subprojects, which may actually be
useful to package. For now we are only bothering packaging frida-gum,
for a planned packaging of https://github.com/lf-/clipper.