nixos/security/pam: add authorizedKeysFiles option for pam_ssh_agent_auth
The current default allows privilege escalation because a process
running as a user can add its own key to the user-writable
authorized_keys file.

This adds an option to configure the authorized keys files for
pam_ssh_agent_auth separately from openssh. The default is only the
/etc/ssh/authorized_keys.d/%u path which is not user-writable.

Since existing users may rely on keys in user-writable files for access
to their system, the option has been renamed so that users are prompted
with upgrade instructions.

Fixes #31611