Don't require ellipsis in NixOS module definitions.
Having used `callPackage` a bit, it seems a bit odd and inconsistent that modules
require `...` in their signatures as opposed to figuring out the
signature and passing only what's requested, like `callPackage` does.

Now I don't know if there is an actual reason for why it is important
for modules to require `...`, but I could not find any and upon a brief
chat with @infinisil, it seemed like there might not be a reason. So if
you know of a reason, please let me know.

Worth noting that, as currently implemented, it's is a breaking change,
namely modules that use at-pattern arg capture don't work (`{ ...
}@args`, args used to contain all available module args, but is now
empty, see newly added failing test for details). I don't have any data
on how common it is to write modules with at-pattern capturing, so would
like to get some guidance on how important it is to preserve this
behavior.

It's possible to make it a non-breaking change by detecting `...` and/or `@args`
with something like #194992 or rather with a new builtin like
was proposed in #7317, and passing all available arguments if detected.

Would like to get your thoughts on the idea to see if there are any
strong objections to pursuing this further.