ruby-modules: Allow the user to specify files the Gemfile depends on
Some Gemfiles depend on the presence of other files. ej. Depends on
.ruby-version to be in the same directory to be able to specify the version from
its contents. This is going to be more common now that bundler supports doing so
using keyword argument, ie. `ruby file: '.ruby-version'`.

For those scenarios, allow bundlerEnv to take a list as an extra argument,
gemfileExtraFiles, to specify a list of files to copy alongside the Gemfile,
lockfile and gemset file. That way when the Gemfile is evaluated it will find
the file it expects. If the file is not found the ruby version is unset and when
bundler tries to validate the ruby version [0] it fails the check and the
process aborts. This happens when trying to run a 'binstubed' executable that
is bundled with a gem such as rackup.

Closes #181693

[0]: https://github.com/rubygems/rubygems/blob/4bbefa9c4db32ef857d92af5ee5e8f98813cfc5e/bundler/lib/bundler/definition.rb#L429