syncterm: Cherry-pick upstream fixes for `pthread_yield` errors
This non-standard function has been deprecated since glibc 2.34 (August 2021),
and calls to it were breaking the build. This has since been fixed in the
upstream Synchronet tree, from which SyncTERM releases are cut, but this was
last done in 2020.

Big thanks to @aexoxea, who not only dug into this, but adapted the upstream
patches to the SyncTERM tree and posted almost this exact diff in #174749 -
where it completely slipped my mind until now, sorry!

Fixes #174749