Jellyfin service starts after network-online.target
To those who use Jellyfin's DLNA server, it can happen that the DLNA server starts before we are connected to the LAN.

When this happens, Jellyfin only publishes the DLNA server on the local ports and is not discoverable by devices in the LAN.

In order to fix this, I'm ensuring that Jellyfin starts after we are connected to the network, making it properly discoverable by DLNA clients.

After making this change, Jellyfin's DLNA server is now working as expected on my machine. It used to be consistently undiscoverable.

I verified that this doesn't break anything in situations where the LAN is not available: I disconnected my laptop from the network and rebooted it and Jellyfin started as expected.

This change was informed by reading the suggestion in this article: https://www.freedesktop.org/wiki/Software/systemd/NetworkTarget/

(cherry picked from commit ec9009f8e493a01533bf738b6de44c16c5ea62b1)