spdlog: fix cross-compilation
Currently, catch2 is listed in checkInputs, which means it is not available
when cross-compiling. On the other hand, -DSPDLOG_BUILD_TESTS=ON is passed
unconditionally, causing the build to fail when cross-compiling. This can be
fixed either by moving catch2 to buildInputs or by only building the tests
when doCheck is true. This patch implements the former solution, so the tests
are always built even if they cannot be run.