riemann_c_client: Optional TLS support & other changes
This adjusts riemann_c_client to support building with either wolfSSL or
GnuTLS (disabled by default, to not introduce a new dependency
automatically), and the CLI tool with json-c (enabled by default).

Also splits the outputs to `bin`, `dev`, and `out`, enables parallel
building, and removes a preBuild step that is no longer necessary, and
enables unit tests at build time.

On top of this, also correct the license: it is LGPL3+, *not* GPL3.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>