nixos/virtualisation/qemu-vm: add support for arbitrary firmwares
When doing tests, we test:

* userspace
* bootloader

And now, we will be able to specify the firmware, so we can test also this part of a system.

Currently, we had the capability to test for 3+ firmwares:

- SeaBIOS, built-in in QEMU, nothing to do
- UEFI, via OVMF (EDK2 project), turn on the useEFIBoot knob and adjust the `virtualisation.efi` config space.
- Custom BIOS, e.g. qboot, set `virtualisation.bios`

Now, we will be able to test for 4+ firmwares with the new addition that is U-Boot.

U-Boot is more than a custom BIOS given the fact we can also make use of it in UEFI mode or with Secure Boot.
Therefore, we start a `virtualisation.uboot` config space to take this into account.

On the long run, we should merge the `virtualisation.bios` to make it difficult to represent illegal states, e.g.
custom BIOS != U-Boot && custom BIOS != null.