junicode: 2.003 -> 2.004
Release announcement:
https://github.com/psb1558/Junicode-font/releases/tag/v2.004

From it:

This release includes (mainly in the italic face) revisions of metrics
and kerning of numbers and of kerning around parentheses and
brackets. It also fixes some spacing problems (in the roman face) with
capitals in parentheses (U+1F110-U+1F129) and a few problems with
disordered or wrong components.