davinci-resolve: add studioVariant
Resolve wants to write the license to `./.license` in the drv store
path. This obviously fails as it is read-only. Conveniently this is a
Bubblewrap-based FHS env, so we can bind-mount a reasonable dir from
$HOME to this path instead. Unfortunately, the dir needs to exist before
launching, so it must be manually created presently e.g.

    mkdir -p ~/.local/share/DaVinciResolve/license

At least the error is fairly obvious:

> bwrap: Can't find source path /home/user/.local/share/DaVinciResolve/license: No such file or directory

Without making the license writable, license activation will always give
a generic error—the actual problem was found with strace.