meli: re-introduce at 0.8.1
This package had been dropped in ac019b9 because it required an
insecure version of openssl. This is no longer the case with this
update.

This new version also enables notmuch support by default in upstream,
detecting the presence of the binary purely at runtime. Therefore,
the matching package build switch has been removed.