team-list.nix: nixos-modules -> module-system
The module system isn't about NixOS, so for consistent messaging,
I'd like to rename this.

If some of you would like to be in a team responsible for NixOS'
_use_ of the module system, that would be great, but for consistency
it should be a separate team.

Note that this attribute isn't used anywhere - only generically at
release time for the status update pings.